modified createBanner iOS Api for Unity Plugin

+++ partner-adapter-codes/unity/unity-sampleapp/UnityInMobiPluginProject/Assets/Plugins/iOS/InMobiBanner.h
@@ -26,8 +26,8 @@ typedef enum
 
 -(id)initBannerAd:(InMobiBannerClientRef*) bannerClient
                     placementId:(NSString*) placementId
-                    width:(CGFloat) width
-                    height:(CGFloat) height
+                    width:(int) width
+                    height:(int) height
                     position:(int) position;
 
 //// A reference to the Unity banner client.
